requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.0.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
redis==3.3.11  # https://github.com/andymccurdy/redis-py
celery==4.4.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
mysqlclient
sqlparse
sorl-thumbnail==12.6.2   # 处理缩略图的packeage
# Django
# ------------------------------------------------------------------------------
django==2.2.9  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4  # https://github.com/django-compressor/django-compressor
django-redis==4.11.0  # https://github.com/niwinz/django-redis
django-taggit==1.1.0
django-mdeditor==0.1.17
django-markdownx==3.0.1
django-contrib-comments==1.9.2

# Django REST Framework
# djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
# coreapi==2.3.3  # https://github.com/core-api/python-client
